Binu merge fix

+++ apps/api2/src/controllers/gettingResponse/commentsController.ts
@@ -4,7 +4,9 @@ import bcrypt from "bcrypt";
 
 import { config } from "dotenv";
 import { PrismaClient } from "../../../node_modules/.prisma/client";
-import { createUploadStream } from "../../../../api/src/modules/stream";
+// import { createUploadStream } from "../../../../api/src/modules/stream";
+import { createUploadStream } from "../../modules/stream";
+
 import stream from "stream";
 // import { timeStamp } from "console";
 
@@ -83,8 +85,8 @@ export const fetchComments = async (req: Request, res: Response) => {
 export const createComments = async (req: Request, res: Response) => {
   const { videoId } = req.params;
   const { content, userId, timeStamp, type, parentCommentId } = req.body;
-  console.log(`parentId:${parentCommentId}`);
-  console.log(`type:${type}`);
+  // console.log(`parentId:${parentCommentId}`);
+  // console.log(`type:${type}`);
 
   try {
     const newComment = await prisma.comment.create({
@@ -110,12 +112,12 @@ export const uploadVideoAndCreateComment = async (
   req: Request,
   res: Response
 ) => {
-  console.log("Hello from uploadVideoAndCreateComment backend");
+  // console.log("Hello from uploadVideoAndCreateComment backend");
   // console.log("Request object:", req);
   const { file } = req as any;
   // const { title, description } = req.body as any; // Assuming you are using a middleware to handle file uploads
-  console.log(req.body);
-  console.log(`file:${file}`);
+  // console.log(req.body);
+  // console.log(`file:${file}`);
   // console.log(`title:${title}`);
   // console.log(`description:${description}`);
 
@@ -143,7 +145,7 @@ export const uploadVideoAndCreateComment = async (
     const uploadStream = createUploadStream(originalname);
     stream1.pipe(uploadStream.writeStream);
     const result = await uploadStream.promise;
-    console.log(result);
+    // console.log(result);
     const { ETag, Location, Bucket, Key } = result;
     try {
       // const creatorId = "d68e3f11-bdab-430f-9dc2-54c2c088864d"; // Replace with the actual user ID
@@ -151,10 +153,10 @@ export const uploadVideoAndCreateComment = async (
       const { videoId } = req.params;
       const { userId, timeStamp, typeComment, parentCommentId } =
         req.body as any;
-      console.log(`userId:${userId}`);
+      // console.log(`userId:${userId}`);
       // console.log(req.body);
-      console.log(`parentId:${parentCommentId}`);
-      console.log(typeComment);
+      // console.log(`parentId:${parentCommentId}`);
+      // console.log(typeComment);
 
       const newComment = await prisma.comment.create({
         data: {
@@ -190,19 +192,68 @@ export const uploadVideoAndCreateComment = async (
 };
 
 export const deleteComment = async (req: Request, res: Response) => {
-  console.log("reached delete comment");
+  // console.log("reached delete comment");
 
   const { commentId } = req.params;
 
+  // try {
+  //   // Delete the comment with the provided commentId
+  //   const deletedComment = await prisma.comment.delete({
+  //     where: {
+  //       id: commentId,
+  //     },
+  //   });
+
+  //   res.status(204).send(deletedComment); // Comment successfully deleted
+  // } catch (error) {
+  //   console.error("Error deleting comment:", error);
+  //   res
+  //     .status(500)
+  //     .json({ error: "An error occurred while deleting the comment." });
+  // }
   try {
-    // Delete the comment with the provided commentId
-    const deletedComment = await prisma.comment.delete({
+    // Check if the comment exists
+    const comment = await prisma.comment.findUnique({
       where: {
         id: commentId,
       },
+      include: {
+        replies: true,
+        parentComment: true,
+      },
     });
 
-    res.status(204).send(deletedComment); // Comment successfully deleted
+    if (!comment) {
+      return res.status(404).json({ error: "Comment not found." });
+    }
+
+    // Check if the comment is a parent comment
+    if (!comment.parentCommentId) {
+      // Delete the replies of the parent comment
+      await prisma.comment.deleteMany({
+        where: {
+          parentCommentId: commentId,
+        },
+      });
+
+      // Delete the parent comment
+      await prisma.comment.delete({
+        where: {
+          id: commentId,
+        },
+      });
+
+      res.status(204).send({ commentType: "parent" }); // Parent comment and replies successfully deleted
+    } else {
+      // Delete the reply comment
+      await prisma.comment.delete({
+        where: {
+          id: commentId,
+        },
+      });
+
+      res.status(204).send({ commentType: "reply" }); // Reply comment successfully deleted
+    }
   } catch (error) {
     console.error("Error deleting comment:", error);
     res
